start testing pathfinding stuff
Wrote several experimental functions inside the Robot object to
determine the possible routes. To facilitate this I made it easier to
access Area objects from the tile, not needing the list of Areas inside
the Board object.

Furthermore a PathTree object is added to store the possible paths of
the beam per robot/tank.

Currently non of this pathfinding stuff works.