Added possibility map
Pathfinding extended with a possibilitymap. If the combined possible
routes of a tank eliminate certain mirror placements (i.e. all tanks
pass a certain tile and ) this mirror state is removed from the
possibility array of that tile.

This also opens the possibility to check if adding a tank to the board
yields new mirror information. If not it can be safely ommited to the
player without harming uniqueness of the solution for that particular
board.